// src/api/models/UpdatePostResponse.ts
/**
 * BrandTruth API v1
 * BrandTruth API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { PostDto } from './PostDto';
import { HttpFile } from '../http/http';

export class UpdatePostResponse {
    'post'?: PostDto;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "post",
            "baseName": "post",
            "type": "PostDto",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return UpdatePostResponse.attributeTypeMap;
    }
    
    public constructor() {
    }
}